updated file structure and Readme
