+++ README.md
@@ -1,1 +1,40 @@
-# ShopManager
+# NSS Backend Capstone: "Shop Manager"
+
+### Requirements:
+
+- [x] React - *C.R.U.D and Routing*
+- [x] T-SQL/.Net/C# - *Authentication, Database, and Data Calls*
+- [x] Github - *Version Control and Project Planning*
+- [x] Display comprehension and mastery of technologies from the 12 months of the developer bootcamp
+
+### Technologies Used:
+> - [HTML5](https://developer.mozilla.org/en-US/docs/Web/HTML)
+> - [CSS](https://developer.mozilla.org/en-US/docs/Glossary/CSS), [Sass](https://sass-lang.com/)
+> - [JavaScript](https://developer.mozilla.org/en-US/docs/Glossary/JavaScript), [JSX]()
+> - [Bootstrap](https://getbootstrap.com/), [Reactstrap](https://reactstrap.github.io/), [React-Bootstrap](https://react-bootstrap.github.io/components/dropdowns/)
+> - [C#](https://docs.microsoft.com/en-us/dotnet/csharp/)
+> - [.Net Core](https://docs.microsoft.com/en-us/dotnet/)
+> - [T-SQL](https://docs.microsoft.com/en-us/sql/t-sql/language-reference?view=sql-server-ver15)
+
+### ERD & Figma
+
+> - [Figma](https://www.figma.com/file/uTBCAC4mEWQXvabUnzSQeG/Shop-Manager?node-id=0%3A1)
+
+## ERD
+<img src="https://i.imgur.com/L3o4Yfu.png">
+
+### Description:
+##### "Shop Manager"
+> Shop Manager is a full stack website that I worked on for my backend capstone project. The idea for the project came from my previous experience as an machine shop manager. I realized there was an issue with all the proper information getting passed to all the correct employees. A company can create an account as a manager, from there they can log in and create more users such as employess. They then can create jobs and parts for those jobs to pass out to the employess. The Employees can then log what job they on working on and when they finished it. The manager then can go back to those jobs and check which jobs have been completed.
+
+#### Landing Page for all Users
+<img src="https://i.imgur.com/cBEuZwL.png">
+
+#### Dashboard for Managers
+<img src="https://i.imgur.com/tF3awP8.png">
+
+#### Create Users page
+<img src="https://i.imgur.com/40zITUV.png">
+
+#### View Users page
+<img src="https://i.imgur.com/T0Oey0q.png">
