+++ ShopManager.sln
