--- ShopManager/ShopManager/Models/Company.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace ShopManager.Models
-{
-    public class Company
-    {
-        public int id { get; set; }
-        public string companyName { get; set; }
-    }
-}
