--- ShopManager/ShopManager/Models/Tool.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace ShopManager.Models
-{
-    public class Tool
-    {
-        public int id { get; set; }
-        public int userId { get; set; }
-        public string toolName { get; set; }
-        public string toolType { get; set; }
-        public int slotNumber { get; set; }
-    }
-}
