--- ShopManager/ShopManager/appsettings.json
@@ -1,13 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft": "Warning",
-      "Microsoft.Hosting.Lifetime": "Information"
-    }
-  },
-  "AllowedHosts": "*",
-  "ConnectionStrings": {
-    "ShopManager": "Server=localhost;Database=ShopManager;Trusted_Connection=True;"
-    }
-  }
