+++ ShopManager/appsettings.json
@@ -6,5 +6,8 @@
       "Microsoft.Hosting.Lifetime": "Information"
     }
   },
-  "AllowedHosts": "*"
-}
+  "AllowedHosts": "*",
+  "ConnectionStrings": {
+    "ShopManager": "Server=localhost;Database=ShopManager;Trusted_Connection=True;"
+    }
+  }
