+++ shopmanager.ui/src/components/createJobForm.js
@@ -1,3 +1,4 @@
+import '../styles/createJob.scss';
 import React from 'react';
 import DatePicker from 'react-datepicker';
 import 'react-datepicker/dist/react-datepicker.css';
@@ -115,42 +116,48 @@ class CreateJobForm extends React.Component {
   render() {
     return (
       <>
-      <h1>Add Job</h1>
-      <Form>
+      <div className='job-parent-div'>
+      <Form className='job-form-parent'>
+      <h5 className='job-sign-in'>Create Job</h5>
         <Form.Group as={Row} controlId='formHorizontalEmail'>
-          <Form.Label column sm={2}>
-            Job Name:
+          <Form.Label column sm="3">
+            Job name:
           </Form.Label>
-          <Col sm={10}>
+          <Col sm={8}>
             <Form.Control
+              className='job-form-boxes'
               type='text'
               name='jobName'
               value={this.state.jobName}
               onChange={this.handleChange}
+              placeholder='Job Name'
               required
             />
           </Col>
         </Form.Group>
         <Form.Group as={Row} controlId='formHorizontalEmail'>
-          <Form.Label column sm={2}>
+        <Form.Label column sm="3">
             Customer:
           </Form.Label>
-          <Col sm={10}>
+          <Col sm={8}>
             <Form.Control
+              className='job-form-boxes'
               type='text'
               name='customer'
               value={this.state.customer}
               onChange={this.handleChange}
+              placeholder='Customer'
               required
             />
           </Col>
         </Form.Group>
         <Form.Group as={Row} controlId='formHorizontalEmail'>
-          <Form.Label column sm={2}>
+        <Form.Label column sm="3">
             Budget:
           </Form.Label>
-          <Col sm={10}>
+          <Col sm={8}>
             <Form.Control
+              className='job-form-boxes'
               type='number'
               name='budget'
               value={this.state.budget}
@@ -160,10 +167,10 @@ class CreateJobForm extends React.Component {
           </Col>
         </Form.Group>
         <Form.Group as={Row} controlId='formHorizontalEmail'>
-          <Form.Label column sm={2}>
+          <Form.Label column sm={4}>
             Date Received:
           </Form.Label>
-          <Col sm={10}>
+          <Col sm={6}>
             <DatePicker
               selected={this.state.setDateRec}
               onChange={this.dateChangeOne}
@@ -173,10 +180,10 @@ class CreateJobForm extends React.Component {
           </Col>
         </Form.Group>
         <Form.Group as={Row} controlId='formHorizontalEmail'>
-          <Form.Label column sm={2}>
+          <Form.Label column sm={4}>
             Date Due:
           </Form.Label>
-          <Col sm={10}>
+          <Col sm={6}>
             <DatePicker
               selected={this.state.setDateDue}
               onChange={this.dateChangeTwo}
@@ -187,16 +194,17 @@ class CreateJobForm extends React.Component {
         </Form.Group>
 
         <Form.Group as={Row}>
-          <Col sm={{ span: 10, offset: 2 }}>
-            <Button type='' onClick={this.onFormSubmit}>
+          <Col>
+            <Button className='job-form-button' onClick={this.onFormSubmit}>
               Submit
             </Button>
-            <Button variant='danger' type='' onClick={this.handleCancel}>
+            <Button className='job-form-cancel-button' onClick={this.handleCancel}>
               Cancel
             </Button>
           </Col>
         </Form.Group>
       </Form>
+      </div>
       </>
     );
   }
