+++ shopmanager.ui/src/components/userTableRow.js
@@ -1,3 +1,4 @@
+import '../styles/viewUser.scss';
 import React from 'react';
 import { Button } from 'react-bootstrap';
 
@@ -14,13 +15,13 @@ export default function UserTableRow({
       <td>{user.firstName}</td>
       <td>{user.lastName}</td>
       <td>
-        <Button onClick={() => changeRouteUpdate(user.id)}>Update</Button>
+        <Button className='table-update-button'onClick={() => changeRouteUpdate(user.id)}>Update</Button>
       </td>
       <td>
-        <Button variant='info' onClick={() => changeRouteSingleView(user.id)}>View</Button>
+        <Button className='table-view-button' variant='info' onClick={() => changeRouteSingleView(user.id)}>View</Button>
       </td>
       <td>
-        <Button variant='danger' onClick={() => removeUser(user.id)}>Delete</Button>
+        <Button className='table-delete-button' onClick={() => removeUser(user.id)}>Delete</Button>
       </td>
     </tr>
   );
