+++ shopmanager.ui/src/styles/createJob.scss
@@ -1,31 +1,31 @@
-.single-user-parent-div {
+.job-parent-div {
   display: flex;
   justify-content: center;
   align-items: center;
-  height: 75vh;
+  height: 90vh;
 }
 
-.single-user-form-parent {
+.job-form-parent {
   background-color: rgb(248, 252, 252);
   padding: 30px;
   border-radius: 5px;
   border: rgb(110, 198, 252) solid 4px;
-  height: 60vh;
+  height: 65vh;
   width: 35%;
 }
 
-.single-user-sign-in {
+.job-sign-in {
   color: rgb(110, 198, 252);
-  padding-right: 2%;
+  padding-right: 68%;
   padding-bottom: 2%;
   text-decoration: underline;
 }
 
-.single-user-labels {
+.job-labels {
   padding-right: 15%;
 }
 
-.single-user-form-boxes {
+.job-form-boxes {
   border-radius: 0px;
   border-bottom-style: solid;
   border-color: rgb(110, 198, 252);
@@ -33,8 +33,17 @@
   background-color: rgb(248, 252, 252);
 }
 
+.job-form-button {
+  color: rgb(248, 252, 252);
+  border-radius: 2px;
+  border: rgb(110, 198, 252) solid 2px;
+  background-color: rgb(110, 198, 252);
+  padding-left: 8%;
+  padding-right: 8%;
+  margin: 3px;
+}
 
-.single-user-form-cancel-button {
+.job-form-cancel-button {
   color: rgb(248, 252, 252);
   border-radius: 2px;
   border: rgb(252, 110, 134) solid 2px;
@@ -44,7 +53,7 @@
   margin: 3px;
 }
 
-.single-user-form-cancel-button:hover {
+.job-form-cancel-button:hover {
   color: rgb(248, 252, 252);
   border-radius: 2px;
   border: rgb(243, 40, 74) solid 2px;
@@ -53,4 +62,3 @@
   padding-right: 8%;
   margin: 3px;
 }
-
