+++ shopmanager.ui/src/styles/viewUser.scss
@@ -1,0 +1,114 @@
+.parent-div {
+  display: flex;
+  height: 45vh;
+}
+
+.table-section {
+  width: 95%;
+  height: 55vh;
+}
+
+table {
+  font-family: 'Arial';
+  margin: 25px auto;
+  border-collapse: collapse;
+  background: #f4f4f4;
+  border: 2px solid rgb(110, 198, 252);
+  tr {
+    &:hover {
+      background: #f4f4f4;
+
+      td {
+        color: rgb(23, 122, 184);
+      }
+    }
+  }
+}
+
+th,
+td {
+  color: rgb(46, 82, 104);
+  border: 1px solid rgb(110, 198, 252);
+  padding: 12px 35px;
+  border-collapse: collapse;
+}
+
+th {
+  background: rgb(188, 212, 226);
+  color: #fff;
+  text-transform: uppercase;
+  font-size: 14px;
+  &.last {
+    border-right: none;
+  }
+}
+
+.job-form-button {
+  color: rgb(248, 252, 252);
+  border-radius: 2px;
+  border: rgb(110, 198, 252) solid 2px;
+  background-color: rgb(110, 198, 252);
+  padding-left: 8%;
+  padding-right: 8%;
+  margin: 3px;
+}
+
+.table-delete-button {
+  color: rgb(248, 252, 252);
+  border-radius: 2px;
+  border: rgb(252, 110, 134) solid 2px;
+  background-color: rgb(252, 110, 134);
+  padding-left: 8%;
+  padding-right: 8%;
+  margin: 3px;
+}
+
+.table-delete-button:hover {
+  color: rgb(248, 252, 252);
+  border-radius: 2px;
+  border: rgb(243, 40, 74) solid 2px;
+  background-color: rgb(243, 40, 74);
+  padding-left: 8%;
+  padding-right: 8%;
+  margin: 3px;
+}
+
+.table-view-button {
+  color: rgb(248, 252, 252);
+  border-radius: 2px;
+  border: rgb(100, 214, 171) solid 2px;
+  background-color: rgb(100, 214, 171);
+  padding-left: 8%;
+  padding-right: 8%;
+  margin: 3px;
+}
+
+.table-view-button:hover {
+  color: rgb(248, 252, 252);
+  border-radius: 2px;
+  border: rgb(44, 139, 103) solid 2px;
+  background-color: rgb(55, 150, 113);
+  padding-left: 8%;
+  padding-right: 8%;
+  margin: 3px;
+}
+
+.table-update-button {
+  color: rgb(248, 252, 252);
+  border-radius: 2px;
+  border: rgb(114, 200, 226) solid 2px;
+  background-color: rgb(114, 200, 226);
+  padding-left: 8%;
+  padding-right: 8%;
+  margin: 3px;
+}
+
+.table-update-button:hover {
+  color: rgb(248, 252, 252);
+  border-radius: 2px;
+  border: rgb(56, 152, 182) solid 2px;
+  background-color: rgb(56, 152, 182);
+  padding-left: 8%;
+  padding-right: 8%;
+  margin: 3px;
+}
